pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "synopsis"
version = "0.1.0"
description = "EPUB 분석 및 검색 시스템 - 계층적 TOC 구조를 추출하고 지능적 청킹으로 데이터베이스에 저장"
authors = [
    { name = "Synopsis Team" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
keywords = ["epub", "text-processing", "search", "nlp", "database", "embedding"]

dependencies = [
    "ebooklib>=0.18",
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.2.0",
    "google-generativeai>=0.3.0",
    "beautifulsoup4>=4.12.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/synopsis-team/synopsis"
Repository = "https://github.com/synopsis-team/synopsis"
Documentation = "https://synopsis.readthedocs.io"
"Bug Tracker" = "https://github.com/synopsis-team/synopsis/issues"

[project.scripts]
synopsis = "synopsis.cli:main"
synopsis-setup = "synopsis.cli:setup_cmd"
synopsis-process = "synopsis.cli:process_cmd"
synopsis-search = "synopsis.cli:search_cmd"

[tool.hatch.build.targets.wheel]
packages = ["synopsis"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["synopsis"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "ebooklib.*",
    "pgvector.*",
    "google.generativeai.*",
    "tqdm.*",
]
ignore_missing_imports = true
